Extract Comment and CommentReply into their own component
